server/requirements.txt
click
flask
flask-cors
flask-migrate
flask-restplus
flask-sqlalchemy
pylint
python-jose
pg8000
Werkzeug==0.16.1
